+++ src/client/components/edit_global_pane/edit_global_pane.js
@@ -23,6 +23,16 @@ Template.edit_global_pane.helpers({
 
 Template.edit_global_pane.events({
     'pr-change .pr-edit-global-pane'( event, instance ){
+        // non-reactively update the passed-in roles
+        const $tree = instance.$( event.currentTarget ).find( '.'+this.pr_div );
+        let roles = this.roles.get();
+        roles.global.all = []
+        $tree.jstree( true ).get_checked_descendants( '#' ).every(( id ) => {
+            roles.global.all.push( id.replace( this.pr_prefix, '' ));
+            return true;
+        });
+        roles.global.direct = Roles._filter( roles.global.all );
+        //console.debug( 'roles.global', roles.global );
         instance.$( event.currentTarget ).trigger( 'pr-global-state', { global: Roles.EditPanel.global() });
     }
 });